requirements.txt
onnx
onnxruntime
matplotlib
pillow
torch-pruning
torchcam
python-dotenv
PyQt6
tqdm
opencv-python